% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.fit.all.log.R
\name{exp.fit.all.log}
\alias{exp.fit.all.log}
\title{exp.fit.all.log}
\usage{
\method{exp}{fit.all.log}(filename, skip.time)
}
\arguments{
\item{filename}{filepaths of .mat files which have a "signal"  and "time" field, or if files are in current
working directory, then just list of filenames.}

\item{skip.time}{number of seconds to skip at the beginning for the exponential fit. N=10 improves the fit.}
}
\description{
Function takes a list of calcium imaging files by filenames in the current working directory (.mat files)
and performs an exponential fit to normalize the deltaF/F signal. _This function does not incorporates a linear
term to account for gradual rise in signal, which can alter the sign of the single exponential._
Will output plots showing orignal and fitted values, as well as corrected values. If an object
is assigned, this will be a vector of corrected values
}
\examples{
data <- exp.fit.all.log(files[1], skip.time = 10)
}
