% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AvoidData.R
\name{format_AvoidData}
\alias{format_AvoidData}
\title{format_AvoidData}
\usage{
format_AvoidData(data, day.correct = "OP50", center.data = FALSE,
  min_p = 0.005, ...)
}
\description{
format_AvoidData
}
\examples{
data \%>\% format_AvoidData()
}
