% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AvoidData_median.R
\name{format_AvoidData_median}
\alias{format_AvoidData_median}
\title{format_AvoidData_median}
\usage{
format_AvoidData_median(data, day.correct = "OP50",
  center.data = FALSE, min_p = 0.005, ...)
}
\description{
format_AvoidData_median
}
\examples{
data \%>\% format_AvoidData_median()
}
