% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SOS.R
\name{format_SOS}
\alias{format_SOS}
\title{format_AvoidData}
\usage{
format_SOS(dataset, day_correct = genotype, reorder_levels = NA, ...)
}
\description{
format_AvoidData
}
\examples{
data \%>\% format_SOS()
}
