% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_layers.R
\docType{data}
\name{ggplot_layers}
\alias{ggplot_layers}
\alias{add.scatter}
\alias{add.median}
\alias{add.mean}
\alias{add.quartiles}
\alias{figure.axes}
\alias{add.n.categorical}
\alias{add.n}
\alias{add.Bayes.CI}
\alias{theme_my_classic}
\alias{fun_length}
\alias{theme_my}
\title{ggplot_layers}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 60.}
\usage{
add.scatter()

add.median(width)

add.mean(width)

add.quartiles(width)

figure.axes()

add.n.categorical(dark)

add.n()

add.Bayes.CI()

theme_my_classic

fun_length(x)

theme_my
}
\arguments{
\item{width}{optional width for median lines}
}
\description{
adds layers to ggplots
}
\section{}{
NA
}

\examples{
p <- ggplot(aes(x=genotype, y=pct)
p + add.scatter
}
\keyword{datasets}
