% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ChemoIndex.R
\name{plot_ChemoIndex}
\alias{plot_ChemoIndex}
\title{plot_ChemoIndex}
\usage{
plot_ChemoIndex(data, xvar = strain, palette = "main", n.pos = 0,
  plot.pos = TRUE, ...)
}
\description{
plot_ChemoIndex
}
\examples{
data \%>\% plot_ChemoIndex()
}
