% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_color_themes.R
\docType{data}
\name{plot_color_themes}
\alias{plot_color_themes}
\alias{plotColors}
\alias{plotColorGetter}
\alias{plotColor_palettes}
\alias{paletteGetter}
\alias{scale_color_plot}
\alias{scale_fill_plot}
\title{plot_color_themes}
\format{An object of class \code{character} of length 8.}
\usage{
plotColors

plotColorGetter(...)

plotColor_palettes

paletteGetter(palette = "main", reverse = FALSE, ...)

scale_color_plot(palette = "main", discrete = TRUE, reverse = FALSE,
  drop = drop, ...)

scale_fill_plot(palette = "main", discrete = TRUE, reverse = FALSE,
  drop = drop, ...)
}
\arguments{
\item{...}{character list set up named themes}

\item{palette}{character for palette selection}

\item{reverse}{logical to reverse color orders}

\item{discrete}{discrete or continuous palette}

\item{...}{additional arguments to colorRampPalette (like alpha etc...)}

\item{palette}{character for palette selection}

\item{reverse}{logical to reverse color orders}

\item{...}{additional arguments to discrete_scale or scale_color_gradientn}

\item{palette}{character for palette selection}

\item{discrete}{discrete or continuous palette}

\item{reverse}{logical to reverse color orders}

\item{...}{additional arguments to discrete_scale or scale_color_gradientn}
}
\description{
standard plot color themes

plotColors
makes a list of named colors for the theme

plotColorGetter
use name of index to pull a vector of hex color codes

PlotColor_palettes

paletteGetter

scale_color_plot

scale_fill_plot
}
\section{}{
NA
}

\examples{
p + scale_fill_plot(palette = "main")
p + add.scatter
}
\keyword{datasets}
