% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGCaMP_multi.R
\name{plotGCaMP_multi}
\alias{plotGCaMP_multi}
\title{plotGCaMP_multi}
\usage{
plotGCaMP_multi(FileFilter, genotype, cue, food)
}
\arguments{
\item{FileFilter}{string to search/subset filenames}

\item{genotype}{label the genotype for these data}

\item{cue}{label the stimulus cue.}
}
\description{
Function is a wrapper for exp.fit.all.log.lin which outputs corrected GCaMP signals as well as plots showing
original and corrected signals. Also plots average trace for GCaMP signal. Inputs are matfiles.
The script searches recursively for matfiles from a starter file (can be any placeholder file). Need to exclude
matfiles that are not GCaMP files either using FileFilter, or by putting only relevant files in the folder.
Plan is to update to include stim
time etc...
}
\examples{
data <- plotGCaMP_multi(N2, genotype = N2, cue = octanol)

}
