% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvidenciaChemo-package.R
\docType{package}
\name{ProvidenciaChemo-package}
\alias{ProvidenciaChemo}
\alias{ProvidenciaChemo-package}
\title{ProvidenciaChemo: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
