% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recenter_fittedValues.R
\name{recenter_fittedValues}
\alias{recenter_fittedValues}
\title{recenter_fittedValues}
\usage{
recenter_fittedValues(data, model, day.correct = FALSE,
  transform = "logit", BayesFit = "fitted_draws", ...)
}
\description{
recenter_fittedValues
}
\examples{
data \%>\% recenter_fittedValues
}
