% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_layers.R
\docType{data}
\name{ggplot_layers}
\alias{ggplot_layers}
\alias{add.scatter}
\alias{add.median}
\alias{add.mean}
\alias{add.quartiles}
\alias{figure.axes}
\alias{add.n.categorical}
\alias{add.n}
\alias{add.Bayes.CI}
\alias{theme_my_classic}
\alias{theme_my_ppt}
\alias{fun_length}
\alias{theme_my}
\alias{g_legend}
\title{ggplot_layers}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 60.}
\usage{
add.scatter(yval)

add.median(yval, width, colour = "black")

add.mean(yval, width, colour = black)

add.quartiles(yval, width)

figure.axes(no.x = TRUE, ...)

add.n.categorical(group, dark, ypos = 0, ...)

add.n(xval)

add.Bayes.CI()

theme_my_classic

theme_my_ppt

fun_length(x)

theme_my

g_legend(a.gplot)
}
\arguments{
\item{width}{optional width for median lines}
}
\description{
adds layers to ggplots
}
\section{}{
NA
}

\examples{
p <- ggplot(aes(x=genotype, y=pct)
p + add.scatter
}
\keyword{datasets}
