% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plasticityIndex.R
\name{plot_plasticityIndex}
\alias{plot_plasticityIndex}
\title{plot_plasticityIndex}
\usage{
plot_plasticityIndex(data, xvar = strain, palette = "main", ...)
}
\description{
plot_plasticityIndex
}
\examples{
data \%>\% plot_plasticityIndex()
}
